# hotel-order-ui/src/app/services/models/page-response-category-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CategoryResponse } from '../models/category-response';
export interface PageResponseCategoryResponse {
  content?: Array<CategoryResponse>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}